+++ plugins/avx2/uint/UMESimdVecUint32_1.h
@@ -1164,12 +1164,12 @@ namespace SIMD {
         }
 
         // GATHER
-        inline SIMDVec_u & gather(uint32_t * baseAddr, uint64_t* indices) {
+        inline SIMDVec_u & gather(uint32_t * baseAddr, uint32_t* indices) {
             mVec = baseAddr[indices[0]];
             return *this;
         }
         // MGATHER
-        inline SIMDVec_u & gather(SIMDVecMask<1> const & mask, uint32_t* baseAddr, uint64_t* indices) {
+        inline SIMDVec_u & gather(SIMDVecMask<1> const & mask, uint32_t* baseAddr, uint32_t* indices) {
             if (mask.mMask == true) mVec = baseAddr[indices[0]];
             return *this;
         }
@@ -1184,12 +1184,12 @@ namespace SIMD {
             return *this;
         }
         // SCATTER
-        inline uint32_t* scatter(uint32_t* baseAddr, uint64_t* indices) const {
+        inline uint32_t* scatter(uint32_t* baseAddr, uint32_t* indices) const {
             baseAddr[indices[0]] = mVec;
             return baseAddr;
         }
         // MSCATTER
-        inline uint32_t*  scatter(SIMDVecMask<1> const & mask, uint32_t* baseAddr, uint64_t* indices) const {
+        inline uint32_t*  scatter(SIMDVecMask<1> const & mask, uint32_t* baseAddr, uint32_t* indices) const {
             if (mask.mMask == true) baseAddr[indices[0]] = mVec;
             return baseAddr;
         }
